Ensure we're passing an int to randint
closes #5861

(cherry picked from commit a6df7e14ed724524d95b53d18a1ef40b83f5ebd5)

+++ pulpcore/tasking/worker.py
@@ -71,7 +71,7 @@ class PulpcoreWorker:
         self.worker = self.handle_worker_heartbeat()
         self.task_grace_timeout = 0
         self.worker_cleanup_countdown = random.randint(
-            WORKER_CLEANUP_INTERVAL / 10, WORKER_CLEANUP_INTERVAL
+            int(WORKER_CLEANUP_INTERVAL / 10), WORKER_CLEANUP_INTERVAL
         )
 
         meter = get_meter(__name__)
